requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
streamlit>=1.30.0
plotly>=5.18.0
